fix: re-organize packages
