--- pkg/infra/tower/README.md
@@ -1,1 +1,0 @@
-<!-- Tower packages -->
